requirements.txt
pandas
apache-airflow
sqlalchemy
psycopg2-binary
apache-airflow-providers-postgres